basic implementation of new preferences window
