+++ CodeEditModules/Modules/AppPreferences/src/Model/FileIconStyle.swift
@@ -1,0 +1,16 @@
+//
+//  FileIconStyle.swift
+//  CodeEdit
+//
+//  Created by Lukas Pistrol on 18.03.22.
+//
+
+import Foundation
+
+public enum FileIconStyle: String, CaseIterable, Hashable {
+    case color
+    case monochrome
+
+    public static let `default` = FileIconStyle.color
+    public static let storageKey = "fileIconStyle"
+}
