+++ CodeEditModules/Modules/AppPreferences/src/PreferencesPlaceholderView.swift
@@ -1,0 +1,22 @@
+//
+//  PreferencesPlaceholderView.swift
+//  
+//
+//  Created by Lukas Pistrol on 30.03.22.
+//
+
+import SwiftUI
+import Preferences
+
+public struct PreferencesPlaceholderView: View {
+
+    public init() {}
+
+    public var body: some View {
+        VStack {
+            Text("Implementation needed")
+        }
+        .frame(width: 844, height: 300)
+        .padding(30)
+    }
+}
